refactor: moved all other sources to lib/
