+++ lib/app.js
@@ -10,7 +10,7 @@ const debug = require('debug')('laser:web')
 const express = require('express')
 const session = require('express-session')
 
-const routes = require('./routes')
+const routes = require('../routes')
 
 const app = express()
