+++ lib/laserAccess.js
@@ -5,7 +5,7 @@ let Gpio
 try {
   Gpio = require('onoff').Gpio
 } catch (_err) {
-  Gpio = require('./test/mock-gpio').Gpio
+  Gpio = require('../test/mock-gpio').Gpio
 }
 
 const EventEmitter = require('events').EventEmitter
