+++ lib/led.js
