+++ lib/mmp.js
