+++ lib/mqtt.js
@@ -1,7 +1,7 @@
 const debug = require('debug')('laser:mqtt')
 const mqtt = require('mqtt')
 
-const config = require('./config')
+const config = require('../config.json')
 
 // MQTT variables
 let maintenanceStatus = 'ok' // Default status is 'ok'
