+++ lib/slack.js
