+++ lib/socket.js
